gitea: fix segmentation fault when showing pull request
When a pull request is submitted on Gitea with the project owners set as
reviewers (not one by one but just a generic everyone) the reviewers
array contains a single NULL which is being dereferenced later.

Fix by replacing NULLs in the retrieved array by a malloced string
stating the project owners as reviewers.

Fixes: https://gitlab.com/herrhotzenplotz/gcli/issues/222
Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>
Signed-off-by: Gavin-John Noonan <mail@gjnoonan.co.uk>