ui: teach gcli to disable the progress spinner
GCLI will display a progress spinner whilst processing, and making
external requests via curl. This can be quite jarring when both not
running in a TTY, or using a dumb terminal. Some people may just wish
to not see it.

gcli has been taught to disable the progress spinner when configured to
do so. This can be configured in one of three ways.

  1) setting GCLI_NOSPINNER to 1/true in the ENV
  2) passing --no-spinner as an argument to the `gcli` command
  3) in your global .gcli. config you can set `disable-spinner=true` in
     the default section

Signed-off-by: Gavin-John Noonan <mail@gjnoonan.co.uk>
Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>