manpages: run a BE spell checker over the manual pages
This makes our manual pages more consistent with the overall rule
to use BE.

There was one exception I made: artefacts (BE) vs. artifacts (AE)
This is because most of the forges use the american spelling of the
word and making it different would involve a) changing the name of
the action in the pipelines subcommand and b) updating the manual
pages. We can possibly add an alias for BE if we really wanted to.

The command I used for spell checking was:

	for f in docs/*.in; do aspell -n --lang=en_GB check $f; done

Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>