comments: add new reply-to option
When specifying -R <comment-id> to the comment subcommand this will
now fetch the comment and put it into the file prefixed with '> ' such
that one can easily reply to it. Currently it only works with issues
and merge requests on GitLab but I will extend it for other use cases
too.

Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>
Signed-off-by: Gavin-John Noonan <mail@gjnoonan.co.uk>