diffutil: Handle diff meta lines (starting with '\') properly
We used to get errors parsing diffs when the file diffed was missing
a trailing newline. This gave a diff line like:

	\ No newline at end of file

This caused a hiccup in the parser which made the line a comment
at the end of the diff which makes no sense.

Also add a test case for this specific bug.

Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>
Signed-off-by: Gavin-John Noonan <mail@gjnoonan.co.uk>