Makefile: add phony auto-target to silence output of manpages when nothing has changed
This is a bit of a know-your-make. What it does is instead of
specifying the manpage targets on the cli it uses dependencies to
generate them and exposes an auto-target that regenerates the
Makefile if needed. This way, because the man pages are now
dependencies of the new phony target, they don't get reported when
they are already up-to-date.

Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>