gitconfig: add support for remotes with ssh-urls containing port numbers
I discovered that when we clone out a repository with ssh but the
URL has a port number in it gcli errors out by default when trying
to fetch anything.
This is because it puts the port number into the owner and the
resulting API endpoint is thus always incorrect.

Fix this by checking if anything after the first colon is a digit
and skip until we find a slash.

Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>