gitconfig: Fix {html,ssh}_extractor for nested projects on Gitlab
This fixes a bug where the review action for pulls would display
bad defaults leading to a miserable UX where the user must correct
the remote and the owner and target branch of the MR.

Fix by scanning the remote-url backwards for slashes thus extracting
the correct owner and repo paths.

This apparently also fixes a couple of bugs related to submodules
such as listing pull requests in submodules.

Related: https://gitlab.com/herrhotzenplotz/gcli/-/issues/241
Tested-by: Bence Ferdinandy <bence@ferdinandy.com>
Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>