cmdconfig: Avoid running into documentation bug by adding alias for api-base
We had a documentation bug and inconsistency regarding the api-base
config value in the configuration file of gcli.

Essentially the tool was searching for apibase where the documentation
said it was looking for api-base (without a dash).

Add an alias and only document api-base as a default. The latter
is to not break existing configurations.

Issue-tracker: https://github.com/herrhotzenplotz/gcli/issues/138
Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>